+++ src/components/SubscriptionManager/index.js
@@ -466,7 +466,7 @@ const SubscriptionSection = ({
         flex={ !isSmallScreen && showCancelCTA ? { grow: 1 } : 'shrink' }
         style={{ height: isSmallScreen ? "fit-content" : "100%" }}
       >
-        <Experiment id="bFcUbpJZS-aZjr5QmZ9JTg">
+        <Experiment id="N0zOeDdQRqCDZb_UEem9Dw">
           <Variant id="0">
             <VariantA
               plans={plansData}