+++ src/routes/habitat/components/CardTabs/tabs/index.js
@@ -1,20 +1,10 @@
 import { h } from 'preact';
-import { useCallback, useEffect } from 'preact/hooks';
+import { useCallback, useEffect, useRef } from 'preact/hooks';
 import { connect } from 'react-redux';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  faInfoCircle,
-  faPaw,
-  faCalendarStar,
-  faPuzzlePiece,
-  faUsers,
-} from '@fortawesome/pro-regular-svg-icons';
-import { faPhotoVideo, faBullhorn } from '@fortawesome/pro-solid-svg-icons';
 import { buildURL, post } from 'Shared/fetch';
-import classnames from 'classnames';
-
-import ClickMessageTip from 'Components/ClickMessageTip';
 import { logGAEvent } from 'Shared/ga';
+import List from 'Components/List';
+import ClickMessageTip from 'Components/ClickMessageTip';
 
 import { updateUserProperty } from '../../../../../redux/actions';
 import { useIsHabitatTabbed } from '../../../../../hooks';
@@ -29,29 +19,34 @@ import {
   QUESTIONS,
 } from '../constants';
 
-import ToggleButton from '../toggleButton';
+import AlbumIcon from './icons/Album.svg';
+import AnimalInfoIcon from './icons/AnimalInfo.svg';
+import BodyIcon from './icons/Body.svg';
+import CalendarIcon from './icons/Calendar.svg';
+import FamilyIcon from './icons/Family.svg';
+import QuestionsIcon from './icons/Question.svg';
+import QuizIcon from './icons/Quiz.svg';
+
+import Tab from './Tab/Tab';
 
 import style from './style.scss';
 
 const Tabs = ({
-  active,
+  activeTab,
   habitatId,
   isAlbumClicked,
   isStreamClicked,
   setActiveTabAction,
   updateUserPropertyAction,
 }) => {
+  const listRef = useRef();
   const isTabbed = useIsHabitatTabbed();
-  const onClick = useCallback(({ target }) => {
-    logGAEvent(
-      'Tab-Navigation-Desktop',
-      target.dataset.value,
-      habitatId,
-    );
+  const onClick = useCallback((tab) => () => {
+    logGAEvent('Tab-Navigation-Desktop', tab, habitatId);
 
-    setActiveTabAction(target.dataset.value);
+    setActiveTabAction(tab);
 
-    if (!isAlbumClicked && target.dataset.value === ALBUM) {
+    if (!isAlbumClicked && tab === ALBUM) {
       post(buildURL('users/steps'), { step: 'isAlbumClicked', value: true })
         .then((data) => updateUserPropertyAction(data))
         .catch((err) => console.error('Error while updating album click indicator', err));
@@ -62,95 +57,100 @@ const Tabs = ({
   // eslint-disable-next-line react-hooks/exhaustive-deps
   useEffect(() => () => setActiveTabAction(MEET), []);
 
+  useEffect(() => {
+    listRef.current.updateLayout();
+  }, []);
+
   return (
-    <div>
-      <ToggleButton
-        className={style.tabBtn}
-        active={active === MEET}
-        value={MEET}
-        onClick={onClick}
-      >
-        <FontAwesomeIcon className={style.icon} icon={faUsers} />
-        Meet The Family
-      </ToggleButton>
+    <div className={style.tabsContainer}>
+      <List className={style.carousel} ref={listRef}>
+        {!isTabbed && (
+          <Tab
+            active={activeTab === ALBUM}
+            title="Album"
+            description="Updates, photos, clips, past talks, and more"
+            onClick={onClick(ALBUM)}
+            color="#769DFF"
+            iconTip={(
+              <ClickMessageTip
+                align={{ bottom: 'top' }}
+                disable={isAlbumClicked || !isStreamClicked}
+                text="View photos and clips"
+                largeIndicator
+              >
+                <img src={AlbumIcon} alt="" style={{ position: 'relative', zIndex: '1' }} />
+              </ClickMessageTip>
+            )}
+          />
+        )}
 
-      <ToggleButton
-        className={style.tabBtn}
-        active={active === INFO}
-        value={INFO}
-        onClick={onClick}
-      >
-        <FontAwesomeIcon className={style.icon} icon={faInfoCircle} />
-        Animal Info
-      </ToggleButton>
+        <Tab
+          active={activeTab === INFO}
+          title="The Species"
+          description="Learn about their habitat, behaviour, and more"
+          onClick={onClick(INFO)}
+          color="#19705A"
+          icon={AnimalInfoIcon}
+        />
 
-      <ToggleButton
-        className={style.tabBtn}
-        active={active === BODY}
-        value={BODY}
-        onClick={onClick}
-      >
-        <FontAwesomeIcon className={style.icon} icon={faPaw} />
-        The Body
-      </ToggleButton>
+        <Tab
+          active={activeTab === MEET}
+          title="The Family"
+          description="Get to know the members of this habitat"
+          onClick={onClick(MEET)}
+          color="#FF97E5"
+          icon={FamilyIcon}
+        />
 
-      <ToggleButton
-        className={style.tabBtn}
-        active={active === QUIZ}
-        value={QUIZ}
-        onClick={onClick}
-      >
-        <FontAwesomeIcon className={style.icon} icon={faPuzzlePiece} />
-        Quiz
-      </ToggleButton>
+        <Tab
+          active={activeTab === QUESTIONS}
+          title="Q&A"
+          description="Ask questions and start discussions"
+          onClick={onClick(QUESTIONS)}
+          color="#FF6E40"
+          icon={QuestionsIcon}
+        />
 
-      <ToggleButton
-        className={style.tabBtn}
-        active={active === QUESTIONS}
-        value={QUESTIONS}
-        onClick={onClick}
-      >
-        <FontAwesomeIcon className={style.icon} icon={faBullhorn} />
-        Q&A
-      </ToggleButton>
+        <Tab
+          active={activeTab === CALENDAR}
+          title="Schedule"
+          description="See when this habitat is online and event times"
+          onClick={onClick(CALENDAR)}
+          color="#C9D341"
+          icon={CalendarIcon}
+        />
 
-      <ToggleButton
-        className={style.tabBtn}
-        active={active === CALENDAR}
-        value={CALENDAR}
-        onClick={onClick}
-      >
-        <FontAwesomeIcon className={style.icon} icon={faCalendarStar} />
-        Calendar
-      </ToggleButton>
+        <Tab
+          active={activeTab === BODY}
+          title="The Body"
+          description="Explore different parts of the animal body"
+          onClick={onClick(BODY)}
+          color="#F05E64"
+          icon={BodyIcon}
+        />
 
-      {!isTabbed && (
-        <ToggleButton
-          className={classnames(style.tabBtn, { [style.albumClickIndicator]: !isAlbumClicked })}
-          active={active === ALBUM}
-          value={ALBUM}
-          onClick={onClick}
-        >
-          <ClickMessageTip
-            align={{ bottom: 'top' }}
-            disable={isAlbumClicked || !isStreamClicked}
-            text="View photos and clips">
-            <FontAwesomeIcon className={style.icon} icon={faPhotoVideo} />
-          </ClickMessageTip>
-          Album
-        </ToggleButton>
-      )}
+        <Tab
+          active={activeTab === QUIZ}
+          title="Quiz"
+          description="Test your knowledge of the species with a fun quiz"
+          onClick={onClick(QUIZ)}
+          color="#418589"
+          icon={QuizIcon}
+        />
+      </List>
     </div>
   );
 };
 
-export default connect(
-  ({
-    user: { isAlbumClicked, isStreamClicked },
-    habitat: { habitatInfo: { _id: habitatId } },
-  }) => ({ isAlbumClicked, isStreamClicked, habitatId }),
-  {
-    setActiveTabAction: setActiveTab,
-    updateUserPropertyAction: updateUserProperty,
-  },
-)(Tabs);
+export default connect(({
+  user: { isAlbumClicked, isStreamClicked },
+  habitat: { habitatInfo: { _id: habitatId }, cards: { activeTab } },
+}) => ({
+  isAlbumClicked,
+  isStreamClicked,
+  habitatId,
+  activeTab,
+}), {
+  setActiveTabAction: setActiveTab,
+  updateUserPropertyAction: updateUserProperty,
+})(Tabs);