+++ src/routes/habitat/components/Questions/index.js
@@ -2,19 +2,31 @@ import { h } from 'preact';
 import { connect } from 'react-redux';
 
 import useFetch from 'use-http';
-import { API_BASE_URL} from 'Shared/fetch';
-import { useCallback, useState, useEffect } from 'preact/hooks';
+import { API_BASE_URL, post, buildURL} from 'Shared/fetch';
+import {
+  useCallback,
+  useState,
+  useEffect,
+  useContext,
+} from 'preact/hooks';
 import Loader from 'Components/Loader';
 import ErrorModal from 'Components/modals/Error';
+import { GlobalsContext } from 'Shared/context';
 
 import AskQuestion from './AskQuestion';
 import Question from './Question';
+import DeleteMessageModal from './DeleteMessageModal';
+import ReportMessageModal from './ReportMessageModal';
 
 import style from './style.scss';
 
 const Questions = ({ habitatId }) => {
+  const { socket } = useContext(GlobalsContext);
   const [questions, setQuestions] = useState([]);
   const [showError, setShowError] = useState();
+  const [showDeletionModal, setShowDeletionModal] = useState(false);
+  const [showReportModal, setShowReportModal] = useState(false);
+  const [selectedQuestion, setSelectedQuestion] = useState(null);
 
   const {
     response,
@@ -46,6 +58,49 @@ const Questions = ({ habitatId }) => {
     }
   }, [error]);
 
+  useEffect(() => {
+    const onQuestionDeleted = ({ questionId }) => {
+      setQuestions((questions) => questions.filter(({ _id }) => _id !== questionId));
+    };
+
+    socket.on('questionDeleted', onQuestionDeleted);
+    return () => {
+      socket.off('questionDeleted', onQuestionDeleted);
+    };
+  }, [socket]);
+
+  const promptDeletion = (questionId) => {
+    setSelectedQuestion(questionId);
+    setShowDeletionModal(true);
+  };
+
+  const promptReport = (questionId) => {
+    setSelectedQuestion(questionId);
+    setShowReportModal(true);
+  };
+
+  const closeDeleteModalHandler = useCallback(() => {
+    setShowDeletionModal(false);
+    setSelectedQuestion(null);
+  }, [setShowDeletionModal, setSelectedQuestion]);
+
+  const closeReportModalHandler = useCallback(() => {
+    setShowReportModal(false);
+    setSelectedQuestion(null);
+  }, [setShowReportModal, setSelectedQuestion]);
+
+  const reportQuestion = () => {
+    post(buildURL(`questions/${selectedQuestion}/report`))
+      .catch((err) => console.error(err));
+    closeReportModalHandler();
+  }
+
+  const deleteQuestion = () => {
+    post(buildURL(`questions/${selectedQuestion}/delete`))
+      .catch((err) => console.error(err));
+    closeDeleteModalHandler();
+  };
+
   if (loading) {
     return (
       <div className={style.questionsContainer}>
@@ -72,9 +127,20 @@ const Questions = ({ habitatId }) => {
           content={text}
           createdAt={createdAt}
           comments={comments}
+          onReportHandler={promptReport}
+          onDeleteHandler={promptDeletion}
         />
       ))}
       {showError && <ErrorModal onClose={() => setShowError(false)} text="Failed to load Q&A" />}
+      {showDeletionModal && (
+        <DeleteMessageModal onClose={closeDeleteModalHandler} onDelete={deleteQuestion} />
+      )}
+      {showReportModal && (
+        <ReportMessageModal
+          onClose={closeReportModalHandler}
+          onReport={reportQuestion}
+        />
+      )}
     </div>
   );
 };